Clicking 'Picture' now takes a picture. The issue was that neither Genymotion preview nor my webcam actually have a 'focus', and therefore it stayed in the 'WAITING_LOCK' state permanently.
TODO: Figure out a better way to determine if the cameras being used have an autofocus or not.
TODO: The images are being saved in an emulated folder, and only as 'pic.jpg', meaning they get overwritten. Let's use a timestamp as the filename and put it somewhere more logical on the SD card.